fixed bug where mirror position query was getting cast to integer (finder credit to Nico)
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@11874 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/ASITiger/ASIMMirror.cpp
@@ -273,11 +273,11 @@ int CMMirror::GetPosition(double& x, double& y)
    ostringstream command; command.str("");
    command << "W " << axisLetterX_;
    RETURN_ON_MM_ERROR ( hub_->QueryCommandVerify(command.str(),":A") );
-   x = (long) (hub_->ParseAnswerAfterPosition(2)/unitMultX_);
+   x = hub_->ParseAnswerAfterPosition(2)/unitMultX_;
    command.str("");
    command << "W " << axisLetterY_;
    RETURN_ON_MM_ERROR ( hub_->QueryCommandVerify(command.str(),":A") );
-   y = (long) (hub_->ParseAnswerAfterPosition(2)/unitMultY_);
+   y = hub_->ParseAnswerAfterPosition(2)/unitMultY_;
    return DEVICE_OK;
 }
